requirements.txt
tensorflow
keras-gpu
tkinter
